Make PigOperators outputs a list
This changes the output variable to a simple Option[String] which
contains the name of the output relation, if there is one.

Another variable, outputs, is added that is a list of all
PigOperators which read the operators output relation (the one whose
name is in `output`).

To prevent the attributes from getting out of sync, make them private
and provide getters and setters that check for problematic changes.