wip: Convert SplitInto into multiple Filters
Not all tests pass at the moment because SplitIntos `outputs` attribute
is always empty. This means that the `outputs` attribute of the new
`Filter` objects can't be set properly, which makes the operators that
actually read from the `Filter`s not part of the new plan that's built
by Rewriter.processPlan().

To make turning one operator into multiple ones possible, extend
`reconstruct` so it correctly deals with List[PigOperator] objects as
well.