use a plugin-style approach for loading backends
we assume that backend libraries are available on classpath, so that piglet can find the classes easily.
To do so, the start scripts pigs/pigsh/pigf have a BACKEND_DIR variable which points to a directory/file. All files within this directy are added to the classpath.

This allows much simpler adding of backends without having to recompile the application.