FileTools#compileToJar returns Option[Path] containing the path to created jar file for current job
If the jar was successfully created, the path to the jar will be returned. Otherwise, in case of an error, this is None.
This was done because PigCompiler and REPL had to create the path to the jar on their own, although it was already created by the compileToJar method,  which introduced an unnecessary possibility for inconsistency.