export config file to $USER_HOME/.piglet/
The config file may contain values that should be editable by the users. However, if we ship our application as a jar, editing the file is no longer (easily) possible.
Thus, on startup we check if $USER_HOME/.piglet/application.conf exists. If not, we copy the file to this location and load it from there. There is also a new cmdline switch (u, update-config) to force copying the shipped file to the user's home.