Add a backend switch to build file to specify backend for integration tests
When running integration tests, the backends must be configured. Since it might be the case that not all backends are configured one can use this switch to run tests for a specific backend only.
Use with: sbt -Dbackend=spark it:test   (or flink)

This switch is considered only for integration tests