prevent empty lines in generated program
For each operator, we create helper code. However, if the operator does not need helper code,
this creates an empty string to which we appended a newline char. This resulted in many empty lines
at the beginning of the program.

Now, the newline is only added, if the helper code is not empty.