fixes #121 by adding embedded code independently from header
embedded code was emitted using the "header1", which also included the `object {` definition for the program.
However, as we do not want the object for the REPL, header1 was not called in this case, which means no embedded code was emitted.

We know have an extra code template for embedded code which can be called independently.