fix bug when using profiling in combination with join
during inserting the timing operator, we removed the old pipe and added the new pipe to the end of the
inputs/outputs list of an operator. However, the code generator used the position of the pipes to find
correct schema for looking up references.

Thus, we now replace an old pipe with a new one.