implement HTTP request for profiling inside the PerfMonitor.
This also includes the scalaj http library in the generated spark job by adding the --packages option to the sparksubmit call.
This way, Spark will automatically download the library from Maven. This is not a nice solution, but this library is much better than
creating the HTTP request with plain java classes