track number of records that each operator has to process
If profiling is enabled, each operator increments its corresponding accumulator to reflect the number
of records that it is processing. This is implemented currently only for filter, foreach, and group by
operators.

The tracking should be performed inside the actual processing loop, e.g. the filter or map statement, so
that we don't need an extra pass of the data, i.e., an additional map operation