implement determination of materialization points and insert STORE/LOAD appropriately
We collect runtime statistics of programs and use them to decide whether an operator should
be materialized. When an operators is to be materialized, we insert a STORE as an additional
consumer and store a mapping of the Ops lineage ID to the generated file used in STORE.

On the other hand, if we find that we have materialized intermediate results of an Op, we
remove that op (and all its predecessors) and insert a LOAD instead. This Load uses the
mapping to find the correct file to load.

This decision whether an Op should be materialized is based on a cost model, that relies on the
operators costs (runtime) as well as the size (bytes) of processed data as well as the frequency
how often an operator is used (Markov model).