schema class names use hash of their fields instead of counter variable
Names of generated classes where created using a counter variable, i.e. _t1_Tuple , _t2_Tuple, etc

This caused problems for materialization, when a program materialized a _t1_Tuple but in the loading program
the same schema class was named different, e.g., _t3_Tuple.

Therefore, we now create those names based on the hash value of their fields so that schema classes that have
the same fields with the same name and type will be guaranteed to be named the same