fix replacing a materialized op with a new Loader
we previously did not maintain the order of the input pipes of a consumer
of the materialized operator. This would be now problem if there is only
one input pipe. For joins (and similar) however, the ordering matters
as we need to find the correct input schema when generating the Scala code.

We now insert the pipe of new Load operator at the exact posisition of
the old operator.