set lineage string for LOAD
when loading a materialized intermediate result, we now can set the lineage string for that loader
so that subsequent operators have the same lineage ID as without loading the materialized data.

When the operators have the same lineage ID as without loading the mat. data, we find their
previously collected statistics for further analysis